# main/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a document for Tainosystems",
        "termsOfService": "http://tainosystems.com",
        "contact": {
            "name": "Peterson Chery",
            "email": "Petersonchery33@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "192.168.10.146:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/emprunts": {
            "post": {
                "description": "Ajouter un Emprunt pour un client",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Emprunt"
                ],
                "summary": "Ajouter un Emprunt pour un client",
                "parameters": [
                    {
                        "type": "string",
                        "description": "api token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID emprunt",
                        "name": "ID_emprunt",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "project",
                        "name": "Emprunt",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Emprunt"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Emprunt"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/emprunts/sup/{id}": {
            "delete": {
                "description": "Delete emprunt by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Emprunt"
                ],
                "summary": "Delete  emprunt by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "api token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID emprunt",
                        "name": "ID_emprunt",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Emprunt"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/emprunts/up/{id}": {
            "put": {
                "description": "Update an emprunt by his ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Emprunt"
                ],
                "summary": "Update an emprunt by his ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "api token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID emprunt",
                        "name": "ID_emprunt",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "project",
                        "name": "Emprunt",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Emprunt"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Emprunt"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/emprunts/{id}": {
            "get": {
                "description": "Get an emprunt by his ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Emprunt"
                ],
                "summary": "Get an emprunt by his ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "api token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID emprunt",
                        "name": "ID_emprunt",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Emprunt"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/livres": {
            "post": {
                "description": "Add a book",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Livre"
                ],
                "summary": "Add a book",
                "parameters": [
                    {
                        "type": "string",
                        "description": "api token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "project",
                        "name": "Livre",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Livre"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Livre"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/livres/sup/{id}": {
            "delete": {
                "description": "Delete a book by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Livre"
                ],
                "summary": "Delete a book by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "api token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID emprunt",
                        "name": "ID_emprunt",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Emprunt"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/livres/up/{id}": {
            "put": {
                "description": "Update an emprunt by his ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Livre"
                ],
                "summary": "Update an emprunt by his ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "api token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID livre",
                        "name": "ID_livre",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "project",
                        "name": "Livre",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Livre"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Livre"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/livres/{id}": {
            "get": {
                "description": "Get a book by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Livre"
                ],
                "summary": "Get a book by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "api token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID Livre",
                        "name": "ID_Livre",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Livre"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Emprunt": {
            "type": "object",
            "properties": {
                "date_emprunt": {
                    "type": "string"
                },
                "date_retour": {
                    "type": "string"
                },
                "id_emprunt": {
                    "type": "integer"
                },
                "id_livre_emprunt": {
                    "type": "integer"
                },
                "id_tenant": {
                    "type": "integer"
                }
            }
        },
        "model.Livre": {
            "type": "object",
            "properties": {
                "auteur": {
                    "type": "string"
                },
                "description_livre": {
                    "type": "string"
                },
                "id_bibliotheque": {
                    "type": "integer"
                },
                "id_livre": {
                    "type": "integer"
                },
                "titre_livre": {
                    "type": "string"
                }
            }
        },
...
    }
}